:truck: Move datastore scripts
