+++ runtime_config.yml
@@ -28,7 +28,7 @@ training_volumes:
 datastore:
   type: redis
   helper_class: RedisHelper
-  helper_module_path: train_conductor.plugins.redis
+  helper_module_path: train_conductor.datastore.redis
   connection:
     host: localhost
     port: 6379
