+++ train_conductor/interfaces/__init__.py
@@ -12,5 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .database import DatabaseBase
 from .training_info import TrainingStatus
